更新数据库建表语句sql文件；修复排版问题；评分十分制；修复传参问题；修复其他错误

+++ src/com/course/evaluation/po/Course.java
@@ -15,7 +15,7 @@ public class Course {
     private Integer twoStar;
     private Integer oneStar;
     private Integer total;
-    private float score;
+    private double score;
     private String image;
 
     public Course() {
@@ -126,11 +126,11 @@ public class Course {
         this.total = total;
     }
 
-    public float getScore() {
+    public double getScore() {
         return score;
     }
 
-    public void setScore(float score) {
+    public void setScore(double score) {
         this.score = score;
     }
